refactor(wl): introduce active window layout boundary
Extract active-window stack and per-window render layout resolution from surface collection.

- Add `window::layout` to resolve stack render state and per-window render layout data before rendering.
- Keep `collect_active_surfaces` responsible for surface rendering, offscreen cache use, popups, shadows, borders, and badges.
- Preserve existing spawn, overlap, fullscreen, maximize, stack, resize, and render behavior while creating a clearer layout/window seam.
