+++ crates/halley-wl/src/animation/mod.rs
@@ -5,7 +5,7 @@ mod nodes;
 mod surface;
 
 pub(crate) use cluster_tiles::{
-    ClusterTileTracks, cluster_tile_rect_for, cluster_tile_rect_from_field,
+    ClusterTileAnimRect, ClusterTileTracks, cluster_tile_rect_for, cluster_tile_rect_from_field,
     cluster_tile_tracks_animating, retain_live_cluster_tile_tracks, set_cluster_tile_target,
 };
 pub(crate) use curves::{ease_in_out_cubic, ease_out_back};
