+++ crates/halley-wl/src/input/mod.rs
@@ -6,6 +6,6 @@ pub(crate) mod pointer;
 pub(crate) use events::{BackendInputEventData, handle_backend_input_event};
 pub(crate) use keyboard::spawn::spawn_command;
 pub(crate) use pointer::{
-    active_node_screen_rect, active_node_surface_transform_screen_details,
-    active_resize_geometry_screen,
+    ActiveResizeGeometryScreen, active_node_screen_rect,
+    active_node_surface_transform_screen_details, active_resize_geometry_screen,
 };
