+++ crates/halley-wl/src/input/pointer/mod.rs
@@ -7,6 +7,6 @@ pub(crate) mod resize;
 mod screenshot;
 
 pub(crate) use resize::{
-    active_node_screen_rect, active_node_surface_transform_screen_details,
-    active_resize_geometry_screen,
+    ActiveResizeGeometryScreen, active_node_screen_rect,
+    active_node_surface_transform_screen_details, active_resize_geometry_screen,
 };
