+++ crates/halley-wl/src/input/pointer/resize/mod.rs
@@ -9,8 +9,8 @@ mod interaction;
 
 pub(crate) use anim::advance_resize_anim;
 pub(crate) use geometry::{
-    active_node_screen_rect, active_node_surface_transform_screen_details,
-    active_resize_geometry_screen,
+    ActiveResizeGeometryScreen, active_node_screen_rect,
+    active_node_surface_transform_screen_details, active_resize_geometry_screen,
 };
 pub(super) use interaction::{begin_resize, finalize_resize, handle_resize_motion};
